+++ packages/atom.io/react/src/use-o.ts
@@ -1,14 +1,9 @@
 import type { ReadableFamilyToken, ReadableToken } from "atom.io"
-import {
-	findInStore,
-	getFromStore,
-	NotFoundError,
-	seekInStore,
-	subscribeToState,
-} from "atom.io/internal"
+import { getFromStore, subscribeToState } from "atom.io/internal"
 import type { Json } from "atom.io/json"
 import * as React from "react"
 
+import { parseStateOverloads } from "./parse-state-overloads"
 import { StoreContext } from "./store-context"
 
 export function useO<T>(token: ReadableToken<T>): T
@@ -22,25 +17,7 @@ export function useO<T, K extends Json.Serializable>(
 	...params: [ReadableFamilyToken<T, K>, K] | [ReadableToken<T>]
 ): T {
 	const store = React.useContext(StoreContext)
-
-	let token: ReadableToken<any>
-	if (params.length === 2) {
-		const family = params[0]
-		const key = params[1]
-
-		if (store.config.lifespan === `immortal`) {
-			const maybeToken = seekInStore(family, key, store)
-			if (!maybeToken) {
-				throw new NotFoundError(family, key, store)
-			}
-			token = maybeToken
-		} else {
-			token = findInStore(family, key, store)
-		}
-	} else {
-		token = params[0]
-	}
-
+	const token = parseStateOverloads(store, ...params)
 	const id = React.useId()
 	return React.useSyncExternalStore<T>(
 		(dispatch) => subscribeToState(token, dispatch, `use-o:${id}`, store),
